Write durations through a unit-bearing hole, and fix XREADGROUP CLAIM
XREADGROUP passed claimMinIdleTime.Value.TotalMilliseconds - a double - to the
writer, so TimeSpan.FromMilliseconds(1500.5) put "CLAIM 1500.5" on the wire and
the server answered "value is not an integer". Both writers now truncate.

The real fix is the hole rather than the cast. TimeSpan is a BCL type and can
never implement IRespArgument, and nothing implemented IRespFormattableArgument
either, so every site wrote (long)x.TotalMilliseconds by hand - four of them on
the new surface - and the shipped writers passed the double straight through.
RespRequestBuilder.AppendFormatted(TimeSpan, string?) now takes the unit, with
a TimeSpan? twin that writes nothing when null, so $"{idle:ms}" and $"{ttl:s}"
are how a duration is written.

Two properties fall out of that. A unit-less $"{ttl}" does not compile at all -
CS0315, no conversion to IRespArgument - so a duration cannot reach the wire
without someone naming the unit. And an unrecognised unit throws: they are our
tokens, not TimeSpan's, and a duration in the wrong unit is still a valid
command, so the call site is the only place it can be caught.

EX/PX down-levelling is untouched and stays where it was: Expiration truncates
to whole milliseconds and drops to seconds when millis % 1000 == 0, and every
TTL parameter on the new surface is an Expiration rather than a TimeSpan. That
is why there is no {x:s} in src - the {x:ms} holes are all commands whose unit
the protocol fixes.

+++ src/StackExchange.Redis/RedisDatabase.cs
@@ -4564,7 +4564,10 @@ namespace StackExchange.Redis
                 if (claimMinIdleTime.HasValue)
                 {
                     writer.WriteRaw("$5\r\nCLAIM\r\n"u8);
-                    writer.WriteBulkString(claimMinIdleTime.Value.TotalMilliseconds);
+
+                    // whole milliseconds: TotalMilliseconds is a double, and a fractional TimeSpan used to
+                    // put "CLAIM 1500.5" on the wire, which the server rejects as not an integer
+                    writer.WriteBulkString((long)claimMinIdleTime.Value.TotalMilliseconds);
                 }
 
                 writer.WriteRaw("$7\r\nSTREAMS\r\n"u8);
@@ -5383,7 +5386,9 @@ namespace StackExchange.Redis
                 if (claimMinIdleTime.HasValue)
                 {
                     writer.WriteRaw("$5\r\nCLAIM\r\n"u8);
-                    writer.WriteBulkString(claimMinIdleTime.Value.TotalMilliseconds);
+
+                    // whole milliseconds; see the multi-stream writer for why
+                    writer.WriteBulkString((long)claimMinIdleTime.Value.TotalMilliseconds);
                 }
 
                 writer.WriteRaw("$7\r\nSTREAMS\r\n"u8);
